// iam-kimbumjun-com/src/interfaces/i-userdetail-dto.ts
// 상세 정보용
export interface IUserDetailDTO {
    id: string;
    fullName: string;
    email: string;
    emailConfirmed: boolean;
    roles: string[];
    phoneNumber?: string;
    twoFactorEnabled: boolean;
    phoneNumberConformed: boolean;
    accessFailedCount: number;
    avata: string;
}